Use mdbook::utils::new_cmark_parser to preserve tables
Before, we would ignore the Markdown settings from mdBook and parse
the content with default settings. This meant that we did not support
Markdown tables, which are an extension to regular Markdown. The
result was that tables were mangled after the preprocessor had run.

We now use the mdBook utility function to construct the Markdown
parser, thus getting consistent results. This makes tables survive the
parsing.