chore(CI): Refactor release pipeline
This refactors mainly the building step of the release
pipeline to compile multiple binaries for certain targets.

Further, I removed the usage of the `actions-rs/*` actions, as
the github organization is only maintained by one person that
doesn't seem to work on this project any more. The actions in this
organization are already affected of some deprecations that will
break workflows in the next year. So it was required to move to
other actions to setup rust anyways.

actions-rs/toolchain#216

To ensure proper caching I also moved to another popular action in the
rust ecosystem that specifically ensures the caching of cargo and rust.

Additionally to the already existing tag, the release workflow now also
creates a Github Release with the name of the tag and the compiled
binaries for each target platform attached to the release and tag.

As a result, users are now able to download the precompiled binaries of
`mdbook-svgbob` for most popular target platforms and can get started
without going through the compilation process and the installation of
cargo / rust.